moved MeasurementsResults
All the *MeasurementResults classes has been moved to separate file in
 Measurement/Results/
