+++ lnst/RecipeCommon/Perf/Measurements/BaseFlowMeasurement.py
@@ -3,8 +3,7 @@ import textwrap
 from lnst.Controller.RecipeResults import ResultType
 from lnst.RecipeCommon.Perf.Measurements.MeasurementError import MeasurementError
 from lnst.RecipeCommon.Perf.Measurements.BaseMeasurement import BaseMeasurement
-from lnst.RecipeCommon.Perf.Measurements.BaseMeasurement import BaseMeasurementResults
-from lnst.RecipeCommon.Perf.Results import SequentialPerfResult
+from lnst.RecipeCommon.Perf.Measurements.Results import FlowMeasurementResults, AggregatedFlowMeasurementResults
 from lnst.RecipeCommon.Perf.Results import ParallelPerfResult
 
 class Flow(object):
@@ -150,140 +149,6 @@ class NetworkFlowTest(object):
     def client_job(self):
         return self._client_job
 
-class FlowMeasurementResults(BaseMeasurementResults):
-    def __init__(self, measurement, flow, warmup_duration=0):
-        super(FlowMeasurementResults, self).__init__(measurement, warmup_duration)
-        self._flow = flow
-        self._generator_results = None
-        self._generator_cpu_stats = None
-        self._receiver_results = None
-        self._receiver_cpu_stats = None
-
-    @property
-    def flow(self):
-        return self._flow
-
-    @property
-    def generator_results(self) -> ParallelPerfResult:
-        return self._generator_results
-
-    @generator_results.setter
-    def generator_results(self, value):
-        self._generator_results = value
-
-    @property
-    def generator_cpu_stats(self):
-        return self._generator_cpu_stats
-
-    @generator_cpu_stats.setter
-    def generator_cpu_stats(self, value):
-        self._generator_cpu_stats = value
-
-    @property
-    def receiver_results(self) -> ParallelPerfResult:
-        return self._receiver_results
-
-    @receiver_results.setter
-    def receiver_results(self, value):
-        self._receiver_results = value
-
-    @property
-    def receiver_cpu_stats(self):
-        return self._receiver_cpu_stats
-
-    @receiver_cpu_stats.setter
-    def receiver_cpu_stats(self, value):
-        self._receiver_cpu_stats = value
-
-    @property
-    def start_timestamp(self):
-        return min(
-            [
-                self.generator_results.start_timestamp,
-                self.generator_cpu_stats.start_timestamp,
-                self.receiver_results.start_timestamp,
-                self.receiver_cpu_stats.start_timestamp,
-            ]
-        )
-
-    @property
-    def end_timestamp(self):
-        return max(
-            [
-                self.generator_results.end_timestamp,
-                self.generator_cpu_stats.end_timestamp,
-                self.receiver_results.end_timestamp,
-                self.receiver_cpu_stats.end_timestamp,
-            ]
-        )
-
-    @property
-    def warmup_end(self):
-        if self.warmup_duration == 0:
-            return self.start_timestamp
-
-        return max(
-            [
-                parallel[self.warmup_duration - 1].end_timestamp
-                for parallel in (*self.generator_results, *self.receiver_results)
-            ]
-        )
-
-    @property
-    def warmdown_start(self):
-        if self.warmup_duration == 0:
-            return self.end_timestamp
-
-        return min(
-            [
-                parallel[-self.warmup_duration].start_timestamp
-                for parallel in (*self.generator_results, *self.receiver_results)
-            ]
-        )
-
-    def time_slice(self, start, end):
-        result_copy = FlowMeasurementResults(self.measurement, self.flow, warmup_duration=0)
-
-        result_copy.generator_cpu_stats = self.generator_cpu_stats.time_slice(start, end)
-        result_copy.receiver_cpu_stats = self.receiver_cpu_stats.time_slice(start, end)
-
-        result_copy.generator_results = self.generator_results.time_slice(start, end)
-        result_copy.receiver_results = self.receiver_results.time_slice(start, end)
-
-        return result_copy
-
-
-class AggregatedFlowMeasurementResults(FlowMeasurementResults):
-    def __init__(self, measurement, flow):
-        super(FlowMeasurementResults, self).__init__(measurement)
-        self._flow = flow
-        self._generator_results = SequentialPerfResult()
-        self._generator_cpu_stats = SequentialPerfResult()
-        self._receiver_results = SequentialPerfResult()
-        self._receiver_cpu_stats = SequentialPerfResult()
-        self._individual_results = []
-
-    @property
-    def individual_results(self):
-        return self._individual_results
-
-    def add_results(self, results):
-        if results is None:
-            return
-        elif isinstance(results, AggregatedFlowMeasurementResults):
-            self.individual_results.extend(results.individual_results)
-            self.generator_results.extend(results.generator_results)
-            self.generator_cpu_stats.extend(results.generator_cpu_stats)
-            self.receiver_results.extend(results.receiver_results)
-            self.receiver_cpu_stats.extend(results.receiver_cpu_stats)
-        elif isinstance(results, FlowMeasurementResults):
-            self.individual_results.append(results)
-            self.generator_results.append(results.generator_results)
-            self.generator_cpu_stats.append(results.generator_cpu_stats)
-            self.receiver_results.append(results.receiver_results)
-            self.receiver_cpu_stats.append(results.receiver_cpu_stats)
-        else:
-            raise MeasurementError("Adding incorrect results.")
 
 class BaseFlowMeasurement(BaseMeasurement):
     @property
