+++ lnst/RecipeCommon/Perf/Measurements/Results/AggregatedCPUMeasurementResults.py
@@ -1,0 +1,28 @@
+from lnst.RecipeCommon.Perf.Results import SequentialPerfResult
+from lnst.RecipeCommon.Perf.Measurements.Results.CPUMeasurementResults import CPUMeasurementResults
+from lnst.RecipeCommon.Perf.Measurements.MeasurementError import MeasurementError
+
+
+class AggregatedCPUMeasurementResults(CPUMeasurementResults):
+    def __init__(self, measurement, host, cpu):
+        super(AggregatedCPUMeasurementResults, self).__init__(measurement, host, cpu)
+        self._individual_results = []
+
+    @property
+    def individual_results(self):
+        return self._individual_results
+
+    @property
+    def utilization(self):
+        return SequentialPerfResult([i.utilization
+                                     for i in self.individual_results])
+
+    def add_results(self, results):
+        if results is None:
+            return
+        elif isinstance(results, AggregatedCPUMeasurementResults):
+            self.individual_results.extend(results.individual_results)
+        elif isinstance(results, CPUMeasurementResults):
+            self.individual_results.append(results)
+        else:
+            raise MeasurementError("Adding incorrect results.")
